+++ plugins/priority/src/startup.rs
@@ -1,0 +1,32 @@
+use crate::metrics::PrometheusExporterConfig;
+use actix_web::dev::Server;
+use actix_web::{web, App, HttpServer};
+use actix_web_opentelemetry::{PrometheusMetricsHandler, RequestMetrics};
+use opentelemetry_api::global;
+use std::net::TcpListener;
+use tracing_actix_web::TracingLogger;
+
+/// Configures and starts the HttpServer
+pub async fn run(
+    listener: TcpListener,
+    request_metrics: PrometheusExporterConfig,
+) -> Result<Server, std::io::Error> {
+    global::set_meter_provider(request_metrics.provider);
+
+    let server = HttpServer::new(move || {
+        App::new()
+            // Logging middleware
+            .wrap(TracingLogger::default())
+            .wrap(RequestMetrics::default())
+            .route(
+                "/metrics",
+                web::get().to(PrometheusMetricsHandler::new(
+                    request_metrics.prom_registry.clone(),
+                )),
+            )
+    })
+    .listen(listener)?
+    .run();
+
+    Ok(server)
+}